Changes Changes.

+++ src/main/java/ru/mail/polis/service/valaubr/HttpService.java
@@ -1,59 +1,86 @@
 package ru.mail.polis.service.valaubr;
 
 import com.google.common.base.Charsets;
-import one.nio.http.HttpServer;
-import one.nio.http.HttpServerConfig;
-import one.nio.http.HttpSession;
-import one.nio.http.Param;
-import one.nio.http.Path;
-import one.nio.http.Request;
-import one.nio.http.RequestMethod;
-import one.nio.http.Response;
+import com.google.common.util.concurrent.ThreadFactoryBuilder;
+import one.nio.http.*;
+import one.nio.net.ConnectionString;
+import one.nio.pool.PoolException;
 import one.nio.server.AcceptorConfig;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.mail.polis.dao.DAO;
 import ru.mail.polis.service.Service;
+import ru.mail.polis.service.valaubr.topology.ModularTopology;
+import ru.mail.polis.service.valaubr.topology.Topology;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
+import java.util.concurrent.*;
 
 public class HttpService extends HttpServer implements Service {
+    private static final String UNIVERSAL_MESSAGE = "Empty id && response is dropped";
     private final DAO dao;
     private final Logger logger = LoggerFactory.getLogger(HttpService.class);
     private final ExecutorService executor;
-    private static final String UNIVERSAL_MESSAGE = "Empty id && response is dropped";
+    private final Map<String, HttpClient> nodeToClient;
+    private final Topology<String> topology;
 
     /**
      * Constructor of the service.
      *
-     * @param port - port of connection
-     * @param base - object of storage
+     * @param port            - port of connection
+     * @param base            - object of storage
+     * @param modularTopology - topology of service
      * @throws IOException - exceptions
      */
     public HttpService(final int port,
                        @NotNull final DAO base,
                        final int threadPool,
-                       final int queueSize) throws IOException {
-        super(config(port));
+                       final int queueSize,
+                       @NotNull final ModularTopology modularTopology) throws IOException {
+        super(config(port, threadPool));
         dao = base;
-        executor = new ThreadPoolExecutor(threadPool, threadPool, 0L,
-                TimeUnit.MILLISECONDS, new ArrayBlockingQueue<>(queueSize));
+        topology = modularTopology;
+        nodeToClient = new HashMap<>();
+
+        for (final String node : topology.all()) {
+            if (topology.isMe(node)) {
+                continue;
+            }
+            final HttpClient client = new HttpClient(new ConnectionString(node + "?timeout=1000"));
+            if (nodeToClient.put(node, client) != null) {
+                throw new IllegalStateException("Duplicate node");
+            }
+        }
+
+        this.executor = new ThreadPoolExecutor(
+                threadPool,
+                queueSize,
+                0L, TimeUnit.MILLISECONDS,
+                new ArrayBlockingQueue<>(queueSize),
+                new ThreadFactoryBuilder()
+                        .setNameFormat("Worker-%d")
+                        .setUncaughtExceptionHandler((thread, e) -> logger.error("error in {} thread", thread, e))
+                        .build(),
+                new ThreadPoolExecutor.AbortPolicy()
+        );
+
     }
 
-    private static HttpServerConfig config(final int port) {
+    private static HttpServerConfig config(final int port, final int threadPool) {
         final AcceptorConfig acceptorConfig = new AcceptorConfig();
         acceptorConfig.port = port;
         acceptorConfig.deferAccept = true;
         acceptorConfig.reusePort = true;
         final HttpServerConfig httpServerConfig = new HttpServerConfig();
         httpServerConfig.acceptors = new AcceptorConfig[]{acceptorConfig};
+        httpServerConfig.selectors = threadPool;
+        httpServerConfig.maxWorkers = threadPool;
+        httpServerConfig.minWorkers = threadPool;
         return httpServerConfig;
     }
 
@@ -67,7 +94,6 @@ public class HttpService extends HttpServer implements Service {
         }
     }
 
-
     /**
      * Return status of server.
      *
@@ -77,7 +103,6 @@ public class HttpService extends HttpServer implements Service {
     public Response status() {
         return Response.ok(Response.OK);
     }
-
 
     /**
      * Getting Entity by id.
@@ -91,37 +116,38 @@ public class HttpService extends HttpServer implements Service {
     @Path("/v0/entity")
     @RequestMethod(Request.METHOD_GET)
     public void get(@Param(required = true, value = "id") @NotNull final String id,
-                    @NotNull final HttpSession session) {
+                    @NotNull final HttpSession session,
+                    @NotNull final Request request) {
+        if (checkId(id, session)) {
+            return;
+        }
         executor.execute(() -> {
-            if (id.strip().isEmpty()) {
+            final ByteBuffer key = ByteBuffer.wrap(id.getBytes(Charsets.UTF_8));
+            final String node = topology.primaryFor(key);
+            if (topology.isMe(node)) {
                 try {
-                    session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
-                    return;
+                    session.sendResponse(Response.ok(
+                            converterFromByteBuffer(dao.get(key))));
+                } catch (NoSuchElementException e) {
+                    logger.error("Record not exist by id = {}", id);
+                    try {
+                        session.sendResponse(new Response(Response.NOT_FOUND, Response.EMPTY));
+                    } catch (IOException ioException) {
+                        logger.error("Record not exist && response is dropped: " + ioException);
+                    }
                 } catch (IOException e) {
-                    logger.error(UNIVERSAL_MESSAGE, e);
+                    logger.error("Error when getting record", e);
+                    try {
+                        session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
+                    } catch (IOException ioException) {
+                        logger.error("Error when getting record && response is dropped: ", ioException);
+                    }
                 }
-            }
-            try {
-                session.sendResponse(Response.ok(
-                        converterFromByteBuffer(dao.get(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8))))));
-            } catch (NoSuchElementException e) {
-                logger.error("Record not exist by id = {}", id);
-                try {
-                    session.sendResponse(new Response(Response.NOT_FOUND, Response.EMPTY));
-                } catch (IOException ioException) {
-                    logger.error("Record not exist && response is dropped: " + ioException);
-                }
-            } catch (IOException e) {
-                logger.error("Error when getting record", e);
-                try {
-                    session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
-                } catch (IOException ioException) {
-                    logger.error("Error when getting record && response is dropped: ", ioException);
-                }
+            } else {
+                proxy(node, request, session);
             }
         });
     }
-
 
     /**
      * Insertion entity dao by id.
@@ -135,25 +161,25 @@ public class HttpService extends HttpServer implements Service {
     @RequestMethod(Request.METHOD_PUT)
     public void put(@Param(required = true, value = "id") @NotNull final String id,
                     @NotNull final Request request, @NotNull final HttpSession session) {
+        if (checkId(id, session)) {
+            return;
+        }
         executor.execute(() -> {
-            if (id.strip().isEmpty()) {
+            final String node = topology.primaryFor(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)));
+            if (topology.isMe(node)) {
                 try {
-                    session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
-                    return;
+                    dao.upsert(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)), ByteBuffer.wrap(request.getBody()));
+                    session.sendResponse(new Response(Response.CREATED, Response.EMPTY));
                 } catch (IOException e) {
-                    logger.error(UNIVERSAL_MESSAGE, e);
+                    logger.error("Error when putting record", e);
+                    try {
+                        session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
+                    } catch (IOException ioException) {
+                        logger.error("Put error && response is dropped:", ioException);
+                    }
                 }
-            }
-            try {
-                dao.upsert(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)), ByteBuffer.wrap(request.getBody()));
-                session.sendResponse(new Response(Response.CREATED, Response.EMPTY));
-            } catch (IOException e) {
-                logger.error("Error when putting record", e);
-                try {
-                    session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
-                } catch (IOException ioException) {
-                    logger.error("Put error && response is dropped:", ioException);
-                }
+            } else {
+                proxy(node, request, session);
             }
         });
     }
@@ -169,32 +195,57 @@ public class HttpService extends HttpServer implements Service {
     @Path("/v0/entity")
     @RequestMethod(Request.METHOD_DELETE)
     public void delete(@Param(required = true, value = "id") @NotNull final String id,
-                       @NotNull final HttpSession session) {
+                       @NotNull final HttpSession session,
+                       @NotNull final Request request) {
+        if (checkId(id, session)) {
+            return;
+        }
         executor.execute(() -> {
-            if (id.strip().isEmpty()) {
+            final String node = topology.primaryFor(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)));
+            if (topology.isMe(node)) {
                 try {
-                    session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
-                    return;
+                    dao.remove(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)));
+                    session.sendResponse(new Response(Response.ACCEPTED, Response.EMPTY));
                 } catch (IOException e) {
-                    logger.error(UNIVERSAL_MESSAGE, e);
+                    logger.error("Error when deleting record", e);
+                    try {
+                        session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
+                    } catch (IOException ioException) {
+                        logger.error("Error when deleting record && response is dropped", e);
+                    }
                 }
-            }
-            try {
-                dao.remove(ByteBuffer.wrap(id.getBytes(Charsets.UTF_8)));
-                session.sendResponse(new Response(Response.ACCEPTED, Response.EMPTY));
-            } catch (IOException e) {
-                logger.error("Error when deleting record", e);
-                try {
-                    session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
-                } catch (IOException ioException) {
-                    logger.error("Error when deleting record && response is dropped", e);
-                }
+            } else {
+                proxy(node, request, session);
             }
         });
     }
 
+    private void proxy(@NotNull String node,
+                       @NotNull Request request,
+                       @NotNull HttpSession session) {
+        try {
+            session.sendResponse(nodeToClient.get(node).invoke(request));
+        } catch (Exception e) {
+            logger.error("Can`t proxy request: ", e);
+        }
+    }
+
+    private boolean checkId(String id, HttpSession session) {
+        if (id.strip().isEmpty()) {
+            try {
+                session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
+                return true;
+            } catch (IOException e) {
+                logger.error(UNIVERSAL_MESSAGE, e);
+                return false;
+            }
+        }
+        return false;
+    }
+
     @Override
-    public void handleDefault(@NotNull final Request request, @NotNull final HttpSession session) throws IOException {
+    public void handleDefault(@NotNull final Request request, @NotNull final HttpSession session) throws
+            IOException {
         executor.execute(() -> {
             try {
                 session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
